assessment schema generator: add generation of basic "required" constraints
this is because i discovered the tests for the schema generator to be using
a now-incorrect declaration fragment as its "definite pass" example, which
could lead to trouble. tests were all behaving themselves because without
"required" constraints it will just ignore missing fields. because of a
single case of a question *not* being required, this is manually overridable
with a "required: False" key in the question's "assessment" section, but
otherwise all questions are assumed to be required - not because it's the
assessment schema's job to enforce that requirement, but it's a stronger
guarantee we keep the content, the tests for the generator and the custom
validation in the supplier frontend in agreement with each other